move pytorch reference to test file. consolidate tests
