+++ vllm/models/qwen4_exp/nvidia/ple_layer.py
@@ -41,7 +41,6 @@ from vllm.v1.attention.backends.short_conv_attn import (
     PleShortConvAttentionBackend,
     PleShortConvAttentionMetadata,
 )
-from vllm.v1.attention.backends.utils import NULL_BLOCK_ID
 
 from ..common.ple import PLEVocabParallelEmbedding
 from .ops.ple_conv import ple_conv
@@ -530,472 +529,6 @@ class Qwen4ExpNGramEmbedding(nn.Module):
         if regular_weights:
             loaded.update(AutoWeightsLoader(self).load_weights(regular_weights))
         return loaded
-
-
-def _short_conv_dilated_decode_pytorch(
-    x_d: torch.Tensor,
-    conv_state: torch.Tensor,
-    conv_weights: torch.Tensor,
-    state_indices_tensor_d: torch.Tensor,
-    has_initial_states_d: torch.Tensor | None,
-    *,
-    conv_state_len: int,
-    dilation: int,
-) -> torch.Tensor:
-    state_indices = state_indices_tensor_d.to(
-        device=conv_state.device, dtype=torch.int64
-    )
-    # FULL cudagraph padded decode rows use NULL_BLOCK_ID. Remap them to
-    # slot 0 for a safe gather, then zero output and skip write-back.
-    valid_state = state_indices != NULL_BLOCK_ID
-    state_indices = torch.where(
-        valid_state, state_indices, torch.zeros_like(state_indices)
-    )
-    if has_initial_states_d is None:
-        has_initial_state = valid_state
-    else:
-        if has_initial_states_d.numel() < state_indices_tensor_d.numel():
-            raise ValueError(
-                "has_initial_states_d size mismatch: "
-                f"got {has_initial_states_d.numel()}, "
-                f"need >= {state_indices_tensor_d.numel()}."
-            )
-        has_initial_state = has_initial_states_d[: state_indices_tensor_d.numel()].to(
-            device=conv_state.device, dtype=torch.bool
-        )
-        has_initial_state = has_initial_state & valid_state
-
-    cached_state = conv_state.index_select(0, state_indices)
-    state = cached_state[..., :conv_state_len].to(x_d.dtype)
-    if conv_state_len > 0:
-        initial_state = torch.where(
-            has_initial_state.view(-1, 1, 1),
-            state,
-            torch.zeros_like(state),
-        )
-        history = torch.cat((initial_state, x_d.unsqueeze(-1)), dim=-1)
-    else:
-        history = x_d.unsqueeze(-1)
-
-    conv_output = F.conv1d(
-        history,
-        conv_weights.unsqueeze(1).contiguous(),
-        groups=history.size(1),
-        dilation=dilation,
-    ).squeeze(-1)
-    output = F.silu(conv_output)
-    output = output * valid_state.view(-1, 1).to(output.dtype)
-
-    if conv_state_len > 0:
-        next_state = history[..., -conv_state_len:]
-        # Padded rows are remapped to the reserved null slot. Preserve its
-        # existing value while writing the new states for valid rows.
-        existing_base_state = cached_state[..., :conv_state_len]
-        safe_next_state = torch.where(
-            valid_state.view(-1, 1, 1),
-            next_state.to(conv_state.dtype),
-            existing_base_state,
-        )
-        cached_state[..., :conv_state_len] = safe_next_state
-        conv_state.index_copy_(0, state_indices, cached_state)
-
-    return output
-
-
-def _short_conv_dilated_prefill_pytorch(
-    x_p: torch.Tensor,
-    metadata: PleShortConvAttentionMetadata,
-    conv_state: torch.Tensor,
-    conv_weights: torch.Tensor,
-    state_indices_tensor_p: torch.Tensor,
-    num_prefills: int,
-    num_decode_tokens: int,
-    num_prefill_tokens: int,
-    *,
-    conv_state_len: int,
-    dilation: int,
-) -> torch.Tensor:
-    # ``non_spec_query_start_loc`` covers the non-spec (decode + prefill)
-    # requests and equals ``query_start_loc`` when spec-decode is inactive.
-    non_spec_query_start_loc = metadata.non_spec_query_start_loc
-    if non_spec_query_start_loc is None:
-        raise ValueError("query_start_loc is required for prefill short-conv")
-    query_start_loc_p = (
-        non_spec_query_start_loc[-num_prefills - 1 :] - num_decode_tokens
-    )
-    # The metadata builder guarantees that the prefill query offsets start
-    # at 0 and end at num_prefill_tokens. Avoid reading those values here,
-    # since doing so would force a device-to-host synchronization.
-    has_initial_states_p = metadata.has_initial_states_p
-    if has_initial_states_p is None:
-        raise ValueError("has_initial_states_p is required for prefill short-conv")
-
-    output = torch.empty_like(x_p)
-    q_starts = query_start_loc_p.to(torch.int64)
-    if state_indices_tensor_p.numel() < num_prefills:
-        raise ValueError(
-            "state_indices_tensor_p size mismatch: "
-            f"got {state_indices_tensor_p.numel()}, "
-            f"need >= {num_prefills}."
-        )
-    if has_initial_states_p.numel() < num_prefills:
-        raise ValueError(
-            "has_initial_states_p size mismatch: "
-            f"got {has_initial_states_p.numel()}, "
-            f"need >= {num_prefills}."
-        )
-    if num_prefills == 0 or x_p.numel() == 0:
-        return output
-    lengths = q_starts[1:] - q_starts[:-1]
-    # Use the CPU-computed packing width from the metadata builder instead
-    # of synchronizing on lengths.max().
-    max_len = metadata.max_prefill_query_len
-    if max_len <= 0:
-        return output
-
-    hidden_size = x_p.shape[1]
-    positions = torch.arange(num_prefill_tokens, device=x_p.device, dtype=torch.int64)
-    req_indices = torch.searchsorted(q_starts[1:], positions, right=True)
-    col_indices = positions - q_starts[req_indices]
-
-    packed_tokens = x_p.new_zeros((num_prefills, max_len, hidden_size))
-    packed_tokens[req_indices, col_indices] = x_p
-    packed_tokens = packed_tokens.transpose(1, 2).contiguous()
-
-    state_indices = state_indices_tensor_p[:num_prefills].to(
-        device=conv_state.device, dtype=torch.int64
-    )
-    valid_state = state_indices != NULL_BLOCK_ID
-    state_indices = torch.where(
-        valid_state, state_indices, torch.zeros_like(state_indices)
-    )
-    has_initial = has_initial_states_p[:num_prefills].to(
-        device=conv_state.device, dtype=torch.bool
-    )
-    if conv_state_len > 0:
-        if conv_state.shape[0] == 0:
-            state = conv_state.new_zeros(
-                (num_prefills, hidden_size, conv_state_len),
-                dtype=x_p.dtype,
-            )
-        else:
-            state = conv_state.index_select(0, state_indices)[..., :conv_state_len].to(
-                x_p.dtype
-            )
-        use_initial_mask = (valid_state & has_initial).view(num_prefills, 1, 1)
-        initial_state = torch.where(
-            use_initial_mask,
-            state,
-            torch.zeros_like(state),
-        )
-        history = torch.cat((initial_state, packed_tokens), dim=-1)
-    else:
-        history = packed_tokens
-
-    conv_output = F.conv1d(
-        history,
-        conv_weights.unsqueeze(1).contiguous(),
-        groups=history.size(1),
-        dilation=dilation,
-    )
-    conv_output = F.silu(conv_output).transpose(1, 2).contiguous()
-
-    token_positions = torch.arange(max_len, device=x_p.device, dtype=torch.int64)
-    valid_tokens = token_positions.view(1, max_len) < lengths.view(num_prefills, 1)
-    valid_output_mask = valid_tokens & valid_state.to(device=x_p.device).view(
-        num_prefills, 1
-    )
-    conv_output.masked_fill_(~valid_output_mask.unsqueeze(-1), 0)
-    output.copy_(conv_output[req_indices, col_indices])
-
-    if conv_state_len > 0 and conv_state.shape[0] > 0:
-        state_starts = lengths.to(device=history.device, dtype=torch.int64).view(
-            num_prefills, 1, 1
-        )
-        state_offsets = torch.arange(
-            conv_state_len, device=history.device, dtype=torch.int64
-        ).view(1, 1, conv_state_len)
-        next_state = history.gather(
-            dim=2,
-            index=(state_starts + state_offsets).expand(-1, history.size(1), -1),
-        )
-        # Write back without a host synchronization. Valid, non-empty rows
-        # receive their new state; padding and zero-length rows keep the
-        # current cache value.
-        existing_state = conv_state.index_select(0, state_indices)
-        existing_base_state = existing_state[..., :conv_state_len]
-        update_mask = valid_state & (lengths.to(device=conv_state.device) > 0)
-        safe_next_state = torch.where(
-            update_mask.view(num_prefills, 1, 1),
-            next_state.to(conv_state.dtype),
-            existing_base_state,
-        )
-        existing_state[..., :conv_state_len] = safe_next_state
-        conv_state.index_copy_(0, state_indices, existing_state)
-    return output
-
-
-def _short_conv_dilated_spec_pytorch(
-    x_spec: torch.Tensor,
-    conv_state: torch.Tensor,
-    conv_weights: torch.Tensor,
-    spec_state_indices_tensor: torch.Tensor,
-    spec_query_start_loc: torch.Tensor,
-    num_accepted_tokens: torch.Tensor,
-    spec_query_len: int,
-    *,
-    conv_state_len: int,
-    dilation: int,
-) -> torch.Tensor:
-    """Dilated short-conv for speculative-decode (MTP) requests.
-
-    Each spec request feeds multiple (draft + 1) query tokens. The conv
-    outputs are computed causally after rolling back the previous draft
-    state by ``num_accepted_tokens - 1``. The current candidate inputs stay
-    in the extended cache for the next forward, matching
-    ``causal_conv1d_update``.
-
-    ``spec_query_len`` (== num_speculative_tokens + 1) is the maximum query
-    length and is a Python int, so no host synchronization is needed; this
-    keeps the path safe for full CUDA-graph capture/replay where the buffers
-    are padded at the request level.
-    """
-    num_reqs = spec_state_indices_tensor.numel()
-    hidden_size = x_spec.size(-1)
-    # Use a fixed packing width instead of synchronizing on lengths.max().
-    max_len = spec_query_len
-    # Full CUDA graphs can pad these buffers. Only the first num_reqs
-    # accepted-token counts belong to actual speculative requests.
-    num_accepted_tokens = num_accepted_tokens[:num_reqs]
-    q_starts = spec_query_start_loc[: num_reqs + 1].to(torch.int64)
-    # Keep the number of real speculative tokens on the device.
-    total_real_tokens = q_starts[num_reqs]
-
-    state_indices = spec_state_indices_tensor.to(
-        device=conv_state.device, dtype=torch.int64
-    )
-    valid_state = state_indices != NULL_BLOCK_ID
-    state_indices = torch.where(
-        valid_state, state_indices, torch.zeros_like(state_indices)
-    )
-    positions = torch.arange(x_spec.size(0), device=x_spec.device, dtype=torch.int64)
-    # Route graph-padded token rows to the discarded dummy request so that
-    # they cannot overwrite real packed data.
-    req_indices = torch.searchsorted(q_starts[1:], positions, right=True)
-    valid_tokens = (positions < total_real_tokens) & (req_indices < num_reqs)
-    clamped_req_indices = req_indices.clamp_max(max(num_reqs - 1, 0))
-    col_indices = (positions - q_starts[clamped_req_indices]).clamp_(0, max_len - 1)
-    pack_req_indices = torch.where(
-        valid_tokens,
-        clamped_req_indices,
-        torch.full_like(req_indices, num_reqs),
-    )
-    pack_col_indices = torch.where(
-        valid_tokens, col_indices, torch.zeros_like(col_indices)
-    )
-
-    # The last request row is the dummy sink for graph padding.
-    packed = x_spec.new_zeros((num_reqs + 1, max_len, hidden_size))
-    packed[pack_req_indices, pack_col_indices] = x_spec
-    packed = packed.transpose(1, 2).contiguous()
-
-    if conv_state_len > 0:
-        cached_state = conv_state.index_select(0, state_indices)
-        rollback_offsets = num_accepted_tokens.to(
-            device=conv_state.device, dtype=torch.int64
-        ).sub(1)
-        rollback_offsets = torch.where(
-            valid_state,
-            rollback_offsets.clamp_(0, max_len - 1),
-            torch.zeros_like(rollback_offsets),
-        )
-        state_offsets = torch.arange(
-            conv_state_len, device=conv_state.device, dtype=torch.int64
-        ).view(1, 1, conv_state_len)
-        rollback_indices = rollback_offsets.view(-1, 1, 1) + state_offsets
-        state = cached_state.gather(2, rollback_indices.expand(-1, hidden_size, -1)).to(
-            x_spec.dtype
-        )
-        state = torch.where(
-            valid_state.view(num_reqs, 1, 1),
-            state,
-            torch.zeros_like(state),
-        )
-        # Append a zeroed dummy-row state to match the [num_reqs + 1] pack.
-        dummy_state = state.new_zeros((1, hidden_size, conv_state_len))
-        state_full = torch.cat((state, dummy_state), dim=0)
-        history = torch.cat((state_full, packed), dim=-1)
-    else:
-        history = packed
-
-    conv_output = F.conv1d(
-        history,
-        conv_weights.unsqueeze(1).contiguous(),
-        groups=history.size(1),
-        dilation=dilation,
-    )
-    conv_output = F.silu(conv_output).transpose(1, 2).contiguous()
-
-    output = conv_output[pack_req_indices, pack_col_indices]
-    output = output * valid_tokens.view(-1, 1).to(output.dtype)
-
-    # Keep all current candidate inputs in the extended state. On the next
-    # target forward, ``num_accepted_tokens - 1`` selects the rollback
-    # window before processing the newly scheduled tokens.
-    if conv_state_len > 0:
-        state_capacity = conv_state_len + max_len - 1
-        if conv_state.size(-1) < state_capacity:
-            raise RuntimeError(
-                "PLE short-conv cache cannot retain speculative tokens: "
-                f"got {conv_state.size(-1)}, need {state_capacity}."
-            )
-        candidate_state = history[:num_reqs, :, 1 : state_capacity + 1]
-        query_lengths = q_starts[1:] - q_starts[:-1]
-        state_positions = torch.arange(
-            state_capacity, device=history.device, dtype=torch.int64
-        ).view(1, 1, state_capacity)
-        update_lengths = (conv_state_len + query_lengths - 1).view(num_reqs, 1, 1)
-        update_mask = valid_state.view(num_reqs, 1, 1) & (
-            state_positions < update_lengths
-        )
-        existing_state = cached_state[..., :state_capacity]
-        next_state = torch.where(
-            update_mask,
-            candidate_state.to(conv_state.dtype),
-            existing_state,
-        )
-        cached_state[..., :state_capacity] = next_state
-        conv_state.index_copy_(0, state_indices, cached_state)
-
-    return output
-
-
-def _short_conv_dilated_dispatch_pytorch(
-    inputs: torch.Tensor,
-    residual: torch.Tensor,
-    metadata: PleShortConvAttentionMetadata,
-    conv_state: torch.Tensor,
-    conv_weights: torch.Tensor,
-    *,
-    conv_state_len: int,
-    dilation: int,
-) -> None:
-    """PyTorch reference for fused short-convolution dispatch."""
-    num_prefills = metadata.num_prefills
-    num_decodes = metadata.num_decodes
-    num_decode_tokens = metadata.num_decode_tokens
-    num_prefill_tokens = metadata.num_prefill_tokens
-    has_prefill = num_prefills > 0
-    has_decode = num_decodes > 0
-    has_spec = metadata.spec_sequence_masks is not None
-    x = inputs[: metadata.num_actual_tokens]
-    residual = residual[: metadata.num_actual_tokens]
-
-    if has_spec:
-        if has_prefill or has_decode:
-            assert metadata.spec_token_indx is not None
-            assert metadata.non_spec_token_indx is not None
-            spec_token_indices = metadata.spec_token_indx.long()
-            non_spec_token_indices = metadata.non_spec_token_indx.long()
-            x_spec = x.index_select(0, spec_token_indices)
-            x_non_spec = x.index_select(0, non_spec_token_indices)
-            residual_spec = residual.index_select(0, spec_token_indices)
-            residual_non_spec = residual.index_select(0, non_spec_token_indices)
-        else:
-            spec_token_indices = None
-            non_spec_token_indices = None
-            x_spec = x
-            x_non_spec = None
-            residual_spec = residual
-            residual_non_spec = None
-    else:
-        spec_token_indices = None
-        non_spec_token_indices = None
-        x_spec = None
-        x_non_spec = x
-        residual_spec = None
-        residual_non_spec = residual
-
-    if has_spec:
-        assert metadata.spec_state_indices_tensor is not None
-        assert metadata.spec_query_start_loc is not None
-        assert metadata.num_accepted_tokens is not None
-        assert x_spec is not None
-        assert residual_spec is not None
-        spec_state_indices = metadata.spec_state_indices_tensor[
-            : metadata.num_spec_decodes
-        ]
-        conv_output = _short_conv_dilated_spec_pytorch(
-            x_spec=x_spec,
-            conv_state=conv_state,
-            conv_weights=conv_weights,
-            spec_state_indices_tensor=spec_state_indices,
-            spec_query_start_loc=metadata.spec_query_start_loc,
-            num_accepted_tokens=metadata.num_accepted_tokens,
-            spec_query_len=metadata.spec_query_len,
-            conv_state_len=conv_state_len,
-            dilation=dilation,
-        )
-        residual_spec.add_(conv_output)
-
-    state_indices = metadata.state_indices_tensor
-    if x_non_spec is not None:
-        assert state_indices is not None
-        assert residual_non_spec is not None
-        if has_prefill:
-            state_indices_d, state_indices_p = torch.split(
-                state_indices, [num_decodes, num_prefills], dim=0
-            )
-            x_d, x_p = torch.split(
-                x_non_spec, [num_decode_tokens, num_prefill_tokens], dim=0
-            )
-            residual_d, residual_p = torch.split(
-                residual_non_spec,
-                [num_decode_tokens, num_prefill_tokens],
-                dim=0,
-            )
-            if has_decode:
-                conv_output = _short_conv_dilated_decode_pytorch(
-                    x_d=x_d,
-                    conv_state=conv_state,
-                    conv_weights=conv_weights,
-                    state_indices_tensor_d=state_indices_d,
-                    has_initial_states_d=metadata.has_initial_states_d,
-                    conv_state_len=conv_state_len,
-                    dilation=dilation,
-                )
-                residual_d.add_(conv_output)
-            conv_output = _short_conv_dilated_prefill_pytorch(
-                x_p=x_p,
-                metadata=metadata,
-                conv_state=conv_state,
-                conv_weights=conv_weights,
-                state_indices_tensor_p=state_indices_p,
-                num_prefills=num_prefills,
-                num_decode_tokens=num_decode_tokens,
-                num_prefill_tokens=num_prefill_tokens,
-                conv_state_len=conv_state_len,
-                dilation=dilation,
-            )
-            residual_p.add_(conv_output)
-        else:
-            conv_output = _short_conv_dilated_decode_pytorch(
-                x_d=x_non_spec,
-                conv_state=conv_state,
-                conv_weights=conv_weights,
-                state_indices_tensor_d=state_indices[: x_non_spec.size(0)],
-                has_initial_states_d=metadata.has_initial_states_d,
-                conv_state_len=conv_state_len,
-                dilation=dilation,
-            )
-            residual_non_spec.add_(conv_output)
-
-    if has_spec and residual_non_spec is not None:
-        assert spec_token_indices is not None
-        assert non_spec_token_indices is not None
-        assert residual_spec is not None
-        residual.index_copy_(0, spec_token_indices, residual_spec)
-        residual.index_copy_(0, non_spec_token_indices, residual_non_spec)
 
 
 class Qwen4ExpPLELayer(nn.Module, MambaBase):
